+++ birdle/admin.py
@@ -1,7 +1,7 @@
 from django.contrib import admin
 from django.contrib.auth.admin import UserAdmin as BaseUserAdmin
 from django.contrib.auth.models import User
-from birds.models import Bird, Game, Guess, Image
+from birdle.models import Bird, Game, Guess, Image
 
 
 @admin.register(Bird)
