+++ birdle/templates/500.html
@@ -1,0 +1,7 @@
+{% extends 'birdle/base.html' %}
+
+{% block content %}
+<div class="d-flex justify-content-center">
+    <img src="{% static 'birdle/dodo.png' %}" alt="A Dodo Bird">
+</div>
+{% endblock %}
