+++ birdle/templates/birdle/base.html
@@ -26,10 +26,10 @@
         <link href="https://cdn.jsdelivr.net/npm/sweetalert2@11.7.3/dist/sweetalert2.min.css" rel="stylesheet">
         <script src="https://cdn.jsdelivr.net/npm/sweetalert2@11"></script>
 
-        <link rel="stylesheet" href="{% static 'birds/style.css' %}">
-        <script src="{% static 'birds/autocomplete.js' %}"></script>
+        <link rel="stylesheet" href="{% static 'birdle/style.css' %}">
+        <script src="{% static 'birdle/autocomplete.js' %}"></script>
 
-        <link rel="shortcut icon" type="image/png" href="{% static 'birds/favicon.png' %}">
+        <link rel="shortcut icon" type="image/png" href="{% static 'birdle/favicon.png' %}">
 
         
   {% block extra_head %}
