+++ birdle/templates/birdle/bird_display.html
@@ -1,3 +1,5 @@
+<!-- bird_display.html -->
+
 <div class="col-sm-8 col-lg-4 mx-auto"> 
     <div id="carouselExampleControls" class="carousel slide" data-bs-ride="carousel" data-bs-theme="dark">
       <div class="carousel-inner">
