+++ birdle/templates/birdle/daily_bird.html
@@ -1,11 +1,11 @@
-<!-- guess.html -->
+<!-- daily_bird.html -->
 
-{% extends 'birds/base.html' %}
+{% extends 'birdle/base.html' %}
 
 {% block content %}
 <div class="container">
 
-  {% include 'birds/bird_display.html'%}
+  {% include 'birdle/bird_display.html'%}
   
   <br>
