+++ birdle/templates/birdle/guess_distribution.html
@@ -1,3 +1,5 @@
+<!-- guess_distribution.html -->
+
 <script src="https://cdn.jsdelivr.net/npm/d3@7"></script>
 <script src="https://cdn.jsdelivr.net/npm/@observablehq/plot@0.6"></script>
