+++ birdle/templates/birdle/info.html
@@ -1,6 +1,6 @@
 <!-- info.html -->
 
-{% extends 'birds/base.html' %}
+{% extends 'birdle/base.html' %}
 
 {% block content %}
 <div class="col-8 mx-auto">
