+++ birdle/templates/birdle/practice.html
@@ -1,4 +1,6 @@
-{% extends 'birds/base.html' %}
+<!-- practice.html -->
+
+{% extends 'birdle/base.html' %}
 
 {% block content %}
 <div class="d-flex justify-content-center">
@@ -19,7 +21,7 @@
 </form>
 
 <br>
-{% include 'birds/bird_display.html'%}
+{% include 'birdle/bird_display.html'%}
 {% comment %} <div class="d-flex justify-content-center">
     <div id="spinner" class="spinner-border" style="width: 3rem; height: 3rem;" role="status">
       <span class="visually-hidden">Loading...</span>
