--- birds/templates/500.html
@@ -1,7 +1,0 @@
-{% extends 'birds/base.html' %}
-
-{% block content %}
-<div class="d-flex justify-content-center">
-    <img src="{% static 'birds/dodo.png' %}" alt="A Dodo Bird">
-</div>
-{% endblock %}
